Rename helpers and remove duplicated code
