+++ src/launch-screens.js
@@ -1,23 +1,9 @@
 /**
- * External dependencies
- */
-const sharp = require('sharp');
-
-/**
  * Internal dependencies
  */
-const { launchScreenSizes } = require('./sizes');
+const { generateFile, launchScreenSizes } = require('./helpers');
 
 /**
  * Generate all app launch screens
  */
-module.exports = (image, folder) => {
-	for (const size of launchScreenSizes) {
-		const [width, height] = size.split('x');
-
-		sharp(image)
-			.resize(Number(width), Number(height))
-			.png()
-			.toFile(`${folder}/launch-screen-${size}.png`);
-	}
-};
+module.exports = (image, folder) => generateFile(image, folder, launchScreenSizes);
